+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/source/preload/BasePreloadManager.java
@@ -17,11 +17,12 @@ package androidx.media3.exoplayer.source.preload;
 
 import static androidx.media3.common.util.Assertions.checkStateNotNull;
 import static androidx.media3.common.util.Util.postOrRun;
+import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
-import static com.google.common.base.Preconditions.checkState;
 
 import android.os.Handler;
 import android.os.Looper;
+import androidx.annotation.CallSuper;
 import androidx.annotation.GuardedBy;
 import androidx.annotation.Nullable;
 import androidx.media3.common.C;
@@ -34,8 +35,8 @@ import androidx.media3.exoplayer.source.MediaSource;
 import com.google.common.base.Predicate;
 import com.google.common.base.Supplier;
 import java.util.Collection;
-import java.util.Comparator;
 import java.util.HashMap;
+import java.util.List;
 import java.util.PriorityQueue;
 
 /**
@@ -49,13 +50,12 @@ public abstract class BasePreloadManager<T, PreloadStatusT> {
 
   /** A base class of the builder of the concrete extension of {@link BasePreloadManager}. */
   protected abstract static class BuilderBase<T, PreloadStatusT> {
-
-    protected final Comparator<T> rankingDataComparator;
     protected final TargetPreloadStatusControl<T, PreloadStatusT> targetPreloadStatusControl;
+    protected RankingDataComparator<T> rankingDataComparator;
     protected Supplier<MediaSource.Factory> mediaSourceFactorySupplier;
 
     public BuilderBase(
-        Comparator<T> rankingDataComparator,
+        RankingDataComparator<T> rankingDataComparator,
         TargetPreloadStatusControl<T, PreloadStatusT> targetPreloadStatusControl,
         Supplier<MediaSource.Factory> mediaSourceFactorySupplier) {
       this.rankingDataComparator = rankingDataComparator;
@@ -67,7 +67,7 @@ public abstract class BasePreloadManager<T, PreloadStatusT> {
   }
 
   private final Object lock;
-  protected final Comparator<T> rankingDataComparator;
+  protected final RankingDataComparator<T> rankingDataComparator;
   private final TargetPreloadStatusControl<T, PreloadStatusT> targetPreloadStatusControl;
   private final MediaSource.Factory mediaSourceFactory;
   private final ListenerSet<PreloadManagerListener> listeners;
@@ -81,8 +81,11 @@ public abstract class BasePreloadManager<T, PreloadStatusT> {
   @Nullable
   private PreloadStatusT targetPreloadStatusOfCurrentPreloadingSource;
 
+  // It's safe to use "this" because the invalidationListener of rankingDataComparator shouldn't be
+  // triggered before exiting this (or subclass's) constructor.
+  @SuppressWarnings("nullness:methodref.receiver.bound")
   protected BasePreloadManager(
-      Comparator<T> rankingDataComparator,
+      RankingDataComparator<T> rankingDataComparator,
       TargetPreloadStatusControl<T, PreloadStatusT> targetPreloadStatusControl,
       MediaSource.Factory mediaSourceFactory) {
     lock = new Object();
@@ -94,6 +97,7 @@ public abstract class BasePreloadManager<T, PreloadStatusT> {
         new ListenerSet<>(applicationHandler.getLooper(), Clock.DEFAULT, (listener, flags) -> {});
     mediaSourceHolderMap = new MediaSourceHolderMap();
     sourceHolderPriorityQueue = new PriorityQueue<>();
+    this.rankingDataComparator.setInvalidationListener(this::invalidate);
   }
 
   /**
@@ -136,6 +140,23 @@ public abstract class BasePreloadManager<T, PreloadStatusT> {
   }
 
   /**
+   * Adds a list of {@linkplain MediaItem media items} with their {@code rankingData} to the preload
+   * manager.
+   *
+   * @param mediaItems The {@linkplain MediaItem media items} to add.
+   * @param rankingDataList The ranking data that are associated with each media item.
+   * @throws IllegalArgumentException If the passed {@code mediaSources} and {@code rankingDataList}
+   *     are different in sizes.
+   */
+  public final void addMediaItems(List<MediaItem> mediaItems, List<T> rankingDataList) {
+    checkArgument(mediaItems.size() == rankingDataList.size());
+    for (int i = 0; i < mediaItems.size(); i++) {
+      add(mediaItems.get(i), rankingDataList.get(i));
+    }
+    invalidate();
+  }
+
+  /**
    * Adds a {@link MediaItem} with its {@code rankingData} to the preload manager.
    *
    * @param mediaItem The {@link MediaItem} to add.
@@ -143,6 +164,23 @@ public abstract class BasePreloadManager<T, PreloadStatusT> {
    */
   public final void add(MediaItem mediaItem, T rankingData) {
     add(mediaSourceFactory.createMediaSource(mediaItem), rankingData);
+  }
+
+  /**
+   * Adds a list of {@linkplain MediaSource media sources} with their {@code rankingData} to the
+   * preload manager.
+   *
+   * @param mediaSources The {@linkplain MediaSource media sources} to add.
+   * @param rankingDataList The ranking data that are associated with each media source.
+   * @throws IllegalArgumentException If the passed {@code mediaSources} and {@code rankingDataList}
+   *     are different in sizes.
+   */
+  public final void addMediaSources(List<MediaSource> mediaSources, List<T> rankingDataList) {
+    checkArgument(mediaSources.size() == rankingDataList.size());
+    for (int i = 0; i < mediaSources.size(); i++) {
+      add(mediaSources.get(i), rankingDataList.get(i));
+    }
+    invalidate();
   }
 
   /**
@@ -194,16 +232,38 @@ public abstract class BasePreloadManager<T, PreloadStatusT> {
    *     {@link MediaItem} and it has been removed, otherwise {@code false}.
    */
   public final boolean remove(MediaItem mediaItem) {
-    if (mediaSourceHolderMap.containsKey(mediaItem)) {
-      MediaSourceHolder mediaHolder = checkNotNull(mediaSourceHolderMap.get(mediaItem));
-      if (isCurrentlyPreloading(mediaHolder)) {
+    @Nullable MediaSourceHolder mediaSourceHolder = mediaSourceHolderMap.get(mediaItem);
+    if (mediaSourceHolder != null) {
+      releaseMediaSourceHolderInternal(mediaSourceHolder);
+      mediaSourceHolderMap.remove(mediaItem);
+      if (isCurrentlyPreloading(mediaSourceHolder)) {
         maybeAdvanceToNextMediaSourceHolder();
       }
-      releaseMediaSourceHolderInternal(mediaHolder);
-      mediaSourceHolderMap.remove(mediaItem);
       return true;
     }
     return false;
+  }
+
+  /**
+   * Removes a list of {@linkplain MediaItem media items} from the preload manager.
+   *
+   * @param mediaItems The {@linkplain MediaItem media items} to remove.
+   */
+  public final void removeMediaItems(List<MediaItem> mediaItems) {
+    for (MediaItem mediaItem : mediaItems) {
+      @Nullable MediaSourceHolder mediaSourceHolder = mediaSourceHolderMap.get(mediaItem);
+      if (mediaSourceHolder != null) {
+        releaseMediaSourceHolderInternal(mediaSourceHolder);
+        mediaSourceHolderMap.remove(mediaItem);
+      }
+    }
+    @Nullable MediaSourceHolder currentMediaSourceHolder;
+    synchronized (lock) {
+      currentMediaSourceHolder = getCurrentlyPreloadingMediaSourceHolder();
+    }
+    if (currentMediaSourceHolder != null && currentMediaSourceHolder.isReleased()) {
+      maybeAdvanceToNextMediaSourceHolder();
+    }
   }
 
   /**
@@ -214,16 +274,38 @@ public abstract class BasePreloadManager<T, PreloadStatusT> {
    *     and it has been removed, otherwise {@code false}.
    */
   public final boolean remove(MediaSource mediaSource) {
-    if (mediaSourceHolderMap.containsKey(mediaSource)) {
-      MediaSourceHolder mediaHolder = checkNotNull(mediaSourceHolderMap.get(mediaSource));
-      if (isCurrentlyPreloading(mediaHolder)) {
+    @Nullable MediaSourceHolder mediaSourceHolder = mediaSourceHolderMap.get(mediaSource);
+    if (mediaSourceHolder != null) {
+      releaseMediaSourceHolderInternal(mediaSourceHolder);
+      mediaSourceHolderMap.remove(mediaSource);
+      if (isCurrentlyPreloading(mediaSourceHolder)) {
         maybeAdvanceToNextMediaSourceHolder();
       }
-      releaseMediaSourceHolderInternal(mediaHolder);
-      mediaSourceHolderMap.remove(mediaSource);
       return true;
     }
     return false;
+  }
+
+  /**
+   * Removes a list of {@linkplain MediaSource media sources} from the preload manager.
+   *
+   * @param mediaSources The {@linkplain MediaSource media sources} to remove.
+   */
+  public final void removeMediaSources(List<MediaSource> mediaSources) {
+    for (MediaSource mediaSource : mediaSources) {
+      @Nullable MediaSourceHolder mediaSourceHolder = mediaSourceHolderMap.get(mediaSource);
+      if (mediaSourceHolder != null) {
+        releaseMediaSourceHolderInternal(mediaSourceHolder);
+        mediaSourceHolderMap.remove(mediaSource);
+      }
+    }
+    @Nullable MediaSourceHolder currentMediaSourceHolder;
+    synchronized (lock) {
+      currentMediaSourceHolder = getCurrentlyPreloadingMediaSourceHolder();
+    }
+    if (currentMediaSourceHolder != null && currentMediaSourceHolder.isReleased()) {
+      maybeAdvanceToNextMediaSourceHolder();
+    }
   }
 
   /**
@@ -447,7 +529,10 @@ public abstract class BasePreloadManager<T, PreloadStatusT> {
    *
    * @param mediaSourceHolder The {@link MediaSourceHolder} to remove.
    */
-  protected abstract void releaseMediaSourceHolderInternal(MediaSourceHolder mediaSourceHolder);
+  @CallSuper
+  protected void releaseMediaSourceHolderInternal(MediaSourceHolder mediaSourceHolder) {
+    mediaSourceHolder.release();
+  }
 
   /** Releases the preload manager, see {@link #release()}. */
   protected void releaseInternal() {}
@@ -463,6 +548,9 @@ public abstract class BasePreloadManager<T, PreloadStatusT> {
   private boolean maybeStartPreloadingNextSourceHolder() {
     if (shouldStartPreloadingNextSource()) {
       MediaSourceHolder preloadingHolder = checkNotNull(sourceHolderPriorityQueue.peek());
+      if (preloadingHolder.isReleased()) {
+        return false;
+      }
       this.targetPreloadStatusOfCurrentPreloadingSource =
           targetPreloadStatusControl.getTargetPreloadStatus(preloadingHolder.rankingData);
       preloadMediaSourceHolderInternal(
@@ -484,11 +572,20 @@ public abstract class BasePreloadManager<T, PreloadStatusT> {
     public final MediaItem mediaItem;
     public final T rankingData;
     private MediaSource mediaSource;
+    private boolean released;
 
     public MediaSourceHolder(MediaItem mediaItem, T rankingData, MediaSource mediaSource) {
       this.mediaItem = mediaItem;
       this.rankingData = rankingData;
       this.mediaSource = mediaSource;
+    }
+
+    public final void release() {
+      released = true;
+    }
+
+    public final boolean isReleased() {
+      return released;
     }
 
     public synchronized MediaSource getMediaSource() {
@@ -498,8 +595,6 @@ public abstract class BasePreloadManager<T, PreloadStatusT> {
     public synchronized void setMediaSource(MediaSource mediaSource) {
       this.mediaSource = mediaSource;
     }
-
-    public void release() {}
 
     @Override
     public int compareTo(MediaSourceHolder o) {
@@ -540,15 +635,6 @@ public abstract class BasePreloadManager<T, PreloadStatusT> {
 
     public synchronized boolean containsKey(MediaItem mediaItem) {
       return mediaItemToMediaSourceHolder.containsKey(mediaItem);
-    }
-
-    public synchronized boolean containsKey(MediaSource mediaSource) {
-      MediaItem mediaItem = mediaSourceToMediaItem.get(mediaSource);
-      if (mediaItem != null) {
-        checkState(mediaItemToMediaSourceHolder.containsKey(mediaItem));
-        return true;
-      }
-      return false;
     }
 
     @Nullable